requirements.txt
opencv-python>=4.8.0
pyzbar>=0.1.9
Pillow>=10.0.0
